Move header to its own directory
